requirements.txt
flask==3.0.2
flask-cors==4.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
requests==2.31.0
PyYAML==6.0.1
fastapi==0.110.0
uvicorn==0.27.1
sse-starlette==1.8.2
pydantic>=2.0.0
websockets==12.0
python-multipart==0.0.9
sounddevice==0.4.6
numpy==1.26.4
openai>=1.0.0
vosk==0.3.45
whisper==1.1.10
pyautogen==0.2.9
asyncio==3.4.3
aiohttp==3.9.3
pytest==8.0.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
structlog>=23.0.0
prometheus-client>=0.17.0
tenacity>=8.2.0